+++ src/Tsufeki/Tenkawa/Php/Reflection/IndexReflectionProvider.php
@@ -4,6 +4,7 @@ namespace Tsufeki\Tenkawa\Php\Reflection;
 
 use Tsufeki\Tenkawa\Php\Reflection\Element\Element;
 use Tsufeki\Tenkawa\Server\Document\Document;
+use Tsufeki\Tenkawa\Server\Document\Project;
 use Tsufeki\Tenkawa\Server\Index\Index;
 use Tsufeki\Tenkawa\Server\Index\IndexEntry;
 use Tsufeki\Tenkawa\Server\Index\Query;
@@ -31,13 +32,12 @@ class IndexReflectionProvider implements ReflectionProvider
     }
 
     /**
+     * @param IndexEntry[] $entries
+     *
      * @resolve Element[]
      */
-    private function search(Query $query, Document $document): \Generator
+    private function process(array $entries): \Generator
     {
-        /** @var IndexEntry[] $entries */
-        $entries = yield $this->index->search($document, $query);
-
         $elements = [];
         foreach ($entries as $entry) {
             $element = $entry->data;
@@ -51,105 +51,206 @@ class IndexReflectionProvider implements ReflectionProvider
         return $elements;
     }
 
-    public function getClass(Document $document, string $fullyQualifiedName): \Generator
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve ClassLike[]
+     */
+    public function getClass($documentOrProject, string $fullyQualifiedName): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_CLASS;
         $query->key = '\\' . ltrim($fullyQualifiedName, '\\');
 
-        return yield $this->search($query, $document);
+        return yield $this->process(yield $this->index->search($documentOrProject, $query));
     }
 
-    public function getFunction(Document $document, string $fullyQualifiedName): \Generator
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve Function_[]
+     */
+    public function getFunction($documentOrProject, string $fullyQualifiedName): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_FUNCTION;
         $query->key = '\\' . ltrim($fullyQualifiedName, '\\');
 
-        return yield $this->search($query, $document);
+        return yield $this->process(yield $this->index->search($documentOrProject, $query));
     }
 
-    public function getConst(Document $document, string $fullyQualifiedName): \Generator
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve Const_[]
+     */
+    public function getConst($documentOrProject, string $fullyQualifiedName): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_CONST;
         $query->key = '\\' . ltrim($fullyQualifiedName, '\\');
 
-        return yield $this->search($query, $document);
+        return yield $this->process(yield $this->index->search($documentOrProject, $query));
     }
 
-    public function getClassesFromUri(Document $document, Uri $uri): \Generator
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve ClassLike[]
+     */
+    public function getClassesFromUri($documentOrProject, Uri $uri): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_CLASS;
         $query->uri = $uri;
 
-        return yield $this->search($query, $document);
+        return yield $this->process(yield $this->index->search($documentOrProject, $query));
     }
 
-    public function getFunctionsFromUri(Document $document, Uri $uri): \Generator
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve Function_[]
+     */
+    public function getFunctionsFromUri($documentOrProject, Uri $uri): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_FUNCTION;
         $query->uri = $uri;
 
-        return yield $this->search($query, $document);
+        return yield $this->process(yield $this->index->search($documentOrProject, $query));
     }
 
-    public function getConstsFromUri(Document $document, Uri $uri): \Generator
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve Const_[]
+     */
+    public function getConstsFromUri($documentOrProject, Uri $uri): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_CONST;
         $query->uri = $uri;
 
-        return yield $this->search($query, $document);
+        return yield $this->process(yield $this->index->search($documentOrProject, $query));
     }
 
-    public function getClassesByShortName(Document $document, string $shortName): \Generator
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve ClassLike[]
+     */
+    public function getClassesByShortName($documentOrProject, string $shortName): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_CLASS;
         $query->key = '\\' . ltrim($shortName, '\\');
         $query->match = Query::SUFFIX;
 
-        return yield $this->search($query, $document);
+        return yield $this->process(yield $this->index->search($documentOrProject, $query));
     }
 
-    public function getFunctionsByShortName(Document $document, string $shortName): \Generator
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve Function_[]
+     */
+    public function getFunctionsByShortName($documentOrProject, string $shortName): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_FUNCTION;
         $query->key = '\\' . ltrim($shortName, '\\');
         $query->match = Query::SUFFIX;
 
-        return yield $this->search($query, $document);
+        return yield $this->process(yield $this->index->search($documentOrProject, $query));
     }
 
-    public function getConstsByShortName(Document $document, string $shortName): \Generator
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve Const_[]
+     */
+    public function getConstsByShortName($documentOrProject, string $shortName): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_CONST;
         $query->key = '\\' . ltrim($shortName, '\\');
         $query->match = Query::SUFFIX;
 
-        return yield $this->search($query, $document);
+        return yield $this->process(yield $this->index->search($documentOrProject, $query));
     }
 
     /**
      * Get class-likes extending, implementing or using (for traits) given class-like.
      *
+     * @param Document|Project $documentOrProject
+     *
      * @resolve string[]
      */
-    public function getInheritingClasses(Document $document, string $fullyQualifiedName): \Generator
+    public function getInheritingClasses($documentOrProject, string $fullyQualifiedName): \Generator
     {
         $query = new Query();
         $query->category = ReflectionIndexDataProvider::CATEGORY_INHERITS;
         $query->key = '\\' . ltrim($fullyQualifiedName, '\\');
 
-        $entries = yield $this->index->search($document, $query);
+        $entries = yield $this->index->search($documentOrProject, $query);
 
         return array_values(array_unique(array_map(function (IndexEntry $entry) {
             return $entry->data;
         }, $entries)));
+    }
+
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve string[]
+     */
+    public function getAllClassNames($documentOrProject): \Generator
+    {
+        $query = new Query();
+        $query->category = ReflectionIndexDataProvider::CATEGORY_CLASS;
+        $query->includeData = false;
+
+        $entries = yield $this->index->search($documentOrProject, $query, true);
+
+        return array_map(function (IndexEntry $entry) {
+            return $entry->key;
+        }, $entries);
+    }
+
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve string[]
+     */
+    public function getAllFunctionNames($documentOrProject): \Generator
+    {
+        $query = new Query();
+        $query->category = ReflectionIndexDataProvider::CATEGORY_FUNCTION;
+        $query->includeData = false;
+
+        $entries = yield $this->index->search($documentOrProject, $query, true);
+
+        return array_map(function (IndexEntry $entry) {
+            return $entry->key;
+        }, $entries);
+    }
+
+    /**
+     * @param Document|Project $documentOrProject
+     *
+     * @resolve string[]
+     */
+    public function getAllConstNames($documentOrProject): \Generator
+    {
+        $query = new Query();
+        $query->category = ReflectionIndexDataProvider::CATEGORY_CONST;
+        $query->includeData = false;
+
+        $entries = yield $this->index->search($documentOrProject, $query, true);
+
+        return array_map(function (IndexEntry $entry) {
+            return $entry->key;
+        }, $entries);
     }
 }